refactor : CrawlerCore 리팩터링 (상수 값 companion object 로 분리, NewsInformation data class로 변경)
