+++ shorts-crawler/src/main/kotlin/com/mashup/shorts/core/HeadLine.kt
@@ -1,0 +1,5 @@
+package com.mashup.shorts.core
+
+enum class HeadLine {
+    HEAD_LINE, NORMAL
+}
